# src/main/java/com/prisma/prismabooking/model/PagedResponse.java
package com.prisma.prismabooking.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class PagedResponse<T> {
    List<T> data;
    String index;
    String elements;
    Long totalElements;
}